+++ cros_gralloc/cros_gralloc_handle.h
@@ -36,7 +36,7 @@ struct cros_gralloc_handle : public native_handle_t {
 	uint32_t magic;
 	uint32_t pixel_stride;
 	int32_t droid_format;
-	int32_t usage; /* Android usage. */
+	int64_t usage; /* Android usage. */
 	uint32_t num_planes;
 	uint64_t reserved_region_size;
 	uint64_t total_size; /* Total allocation size */
